[WIP] First Commit
Latest state of AlgaeCraft for MC 1.7.2 as of July 7, 2014. Still is WIP
stages. Modular structure is generalized, but subject to change. API not
uptodate with other features. All Items are updated and some new Items
are added. Not all blocks are updated. Crafting recipes do not work. New
World type and Biomes WIP. Fish are not yet updated.